Revert "attempt at fixing enum type to ssize_t"
This reverts commit 6cb20b1d4148facfa2b16e4b664ca900ad31ab01.

+++ include/mqtt.h
@@ -214,9 +214,9 @@ struct mqtt_fixed_header {
  * @see mqtt_error_str
  */
 enum MQTTErrors {
-    MQTT_ERROR_UNKNOWN= (ssize_t) INT_MIN,
+    MQTT_ERROR_UNKNOWN=INT_MIN,
     __ALL_MQTT_ERRORS(GENERATE_ENUM)
-    MQTT_OK = (ssize_t) 1
+    MQTT_OK = 1
 };
 
 /**
